mqtt: Handle connection errors
And process deaths.

Connect errors can occur leading (with spawn_link) to terminating this
process. The only way to get a {error, Reason} from the connection is to
trap exits. This then "breaks" supervision if the connection dies other
times.

Trap exits and if it's the connection pid, try again!